Add cygwin tests.
